# backend/openapi.yaml
openapi: 3.1.0
info:
  title: MongoDB Index Manager API
  version: v1
  description: |
    This API provides endpoints for managing MongoDB databases and their indexes. 
    The API uses JWT-based authentication with access tokens and refresh tokens.
    
    **Base Path:** `/api/doctor-manager-api/v1`
    
    **Content-Type:** `application/json`
    
    ## Authentication
    
    The API uses JWT (JSON Web Tokens) for authentication. There are two types of tokens:
    
    - **Access Token**: Used for authenticated requests (short-lived, default: 10 minutes)
    - **Refresh Token**: Used to obtain new access tokens (long-lived, default: 24 hours)
    
    All authenticated requests must include the access token in the `Authorization` header:
    
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Response Format
    
    All responses follow a standard format:
    
    **Success Response:**
    ```json
    {
      "status_code": 200,
      "error_code": 0,
      "data": { ... }
    }
    ```
    
    **Success Response with Pagination:**
    ```json
    {
      "status_code": 200,
      "error_code": 0,
      "data": [ ... ],
      "extra": {
        "limit": 10,
        "page": 1,
        "total": 100
      }
    }
    ```
    
    **Error Response:**
    ```json
    {
      "status_code": 400,
      "error_code": 0,
      "error": "Error message or validation errors"
    }
    ```
    
    ## Error Codes
    
    - `200 OK`: Request successful
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated requests
    refreshTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT refresh token for obtaining new access tokens
  schemas:
...
    SuccessResponseWithPagination:
      type: object
      properties:
        status_code:
          type: integer
          example: 200
        error_code:
          type: integer
          example: 0
        data:
          type: array
          items: { }
        extra:
          $ref: '#/components/schemas/PaginationMetadata'
      required:
        - status_code
        - error_code
        - data
        - extra
    ErrorResponse:
      type: object
      properties:
        status_code:
          type: integer
          example: 400
        error_code:
          type: integer
          example: 0
        error:
          oneOf:
            - type: string
            - type: object
      required:
        - status_code
        - error_code
        - error
    PaginationMetadata:
      type: object
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 50
          example: 10
        page:
          type: integer
          minimum: 1
          example: 1
        total:
          type: integer
          nullable: true
          example: 100
      required:
        - limit
        - page
    ObjectID:
      type: string
      pattern: '^[0-9a-fA-F]{24}$'
      description: MongoDB ObjectID (24-character hexadecimal string)
      example: "507f1f77bcf86cd799439011"
...
    DatabaseUpdateResponse:
      type: object
      properties:
        status_code:
          type: integer
          example: 200
        error_code:
          type: integer
          example: 0
        data:
          type: object
          properties:
            success:
              type: boolean
              example: true
      required:
        - status_code
        - error_code
        - data
...
    DatabaseListCollectionsRequest:
      type: object
      properties:
        database_id:
          $ref: '#/components/schemas/ObjectID'
        query:
          type: string
          nullable: true
        page:
          type: integer
          minimum: 0
          nullable: true
          default: 1
        limit:
          type: integer
          minimum: 0
          nullable: true
          default: 50
          maximum: 50
      required:
        - database_id
    DatabaseCollectionItem:
      type: object
      properties:
        collection:
          type: string
        total_indexes:
          type: integer
    DatabaseListCollectionsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponseWithPagination'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/DatabaseCollectionItem'
    DatabaseCreateCollectionRequest:
      type: object
      properties:
        database_id:
          $ref: '#/components/schemas/ObjectID'
        collection:
          type: string
          description: Collection name to create
      required:
        - database_id
        - collection
    DatabaseCreateCollectionResponse:
      type: object
      properties:
        status_code:
          type: integer
          example: 201
        error_code:
          type: integer
          example: 0
        data:
          type: object
          properties:
            success:
              type: boolean
              example: true
      required:
        - status_code
        - error_code
        - data
    DatabaseUpdateCollectionRequest:
      type: object
      properties:
        database_id:
          $ref: '#/components/schemas/ObjectID'
        collection:
          type: string
        new_collection:
          type: string
      required:
        - database_id
        - collection
        - new_collection
    DatabaseUpdateCollectionResponse:
      $ref: '#/components/schemas/DatabaseUpdateResponse'
    DatabaseDeleteCollectionRequest:
      type: object
      properties:
        database_id:
          $ref: '#/components/schemas/ObjectID'
        collection:
          type: string
      required:
        - database_id
        - collection
    DatabaseDeleteCollectionResponse:
      $ref: '#/components/schemas/DatabaseUpdateResponse'

    # Index Schemas
...
  responses:
    BadRequest:
      description: Bad request - validation errors or invalid data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status_code: 400
            error_code: 0
            error: "Field wrong type"
    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status_code: 401
            error_code: 0
            error: "Token is required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status_code: 404
            error_code: 0
            error: "Resource not found"
...
    PreconditionFailed:
      description: Precondition failed - cannot connect to database or get indexes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status_code: 412
            error_code: 0
            error: "Cannot connect to database"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status_code: 500
            error_code: 0
            error: "Internal server error"
paths:
...
  /databases/collections/list:
    post:
      tags:
        - Database
      summary: List collections
      description: Get a paginated list of collections for a database with index counts
      operationId: listCollections
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseListCollectionsRequest'
      responses:
        '200':
          description: Collections retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseListCollectionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /databases/collections/:
    post:
      tags:
        - Database
      summary: Create a collection
      description: Create a new MongoDB collection in the specified database
      operationId: createCollection
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseCreateCollectionRequest'
      responses:
        '201':
          description: Collection created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseCreateCollectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Database
      summary: Update a collection
      description: Rename a collection in the specified database
      operationId: updateCollection
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseUpdateCollectionRequest'
      responses:
        '200':
          description: Collection updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseUpdateCollectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Database
      summary: Delete a collection
      description: Delete a collection and its indexes from the specified database
      operationId: deleteCollection
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseDeleteCollectionRequest'
      responses:
        '200':
          description: Collection deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseDeleteCollectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Index Endpoints
...